moving post emoji message handler to separate file
